pyproject.toml
[tool.poetry]
name = "pastapull"
version = "0.1.0"
description = ""
authors = ["Max-Eugène Mamet <max-eugene@hotmail.fr>"]
license = "MIT Licence"
readme = "README.md"
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
requests = "^2.32.3"
langchain = "^0.3.0"
openai = "^1.47.1"
langchain-openai = "^0.2.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.35"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
